requirements.txt
httpx==0.23.0
fastapi==0.85.0
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-instrumentation-logging==0.34b0
opentelemetry-exporter-jaeger-thrift==1.13.0
opentelemetry-exporter-jaeger-proto-grpc==1.13.0
opentelemetry-api==1.13.0
opentelemetry-sdk==1.13.0
mariadb==1.0.9
pydantic